SQL-507: Make MongoResultSetMetaData abstract and add subtypes
