+++ src/main/java/com/mongodb/jdbc/MongoSQLResultSetMetaData.java
@@ -1,0 +1,60 @@
+package com.mongodb.jdbc;
+
+import org.bson.BsonType;
+
+import java.sql.ResultSetMetaData;
+import java.sql.SQLException;
+import java.sql.SQLFeatureNotSupportedException;
+import java.sql.Types;
+
+public class MongoSQLResultSetMetaData extends MongoResultSetMetaData implements ResultSetMetaData {
+
+    public MongoSQLResultSetMetaData(MongoResultDoc metadataDoc) {
+        super(metadataDoc);
+    }
+
+    @Override
+    public int getColumnCount() throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public int isNullable(int column) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public String getColumnLabel(int column) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public String getColumnName(int column) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public String getTableName(int column) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public String getCatalogName(int column) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public BsonType getBsonType(int column) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public int getColumnType(int column) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+
+    @Override
+    public String getColumnTypeName(int column) throws SQLException {
+        throw new SQLFeatureNotSupportedException("TODO");
+    }
+}
